MUIPopString.c - Fix a crash if it fails to create the GUI. Types.h - fix definitions to work correctly, when used with older toolchains that pull in pthreads/posix types. various - use ipts/siptr types to correctly use (signed) integers large enough to store pointers.

+++ FryingPan/ISOBuilder/Structure/ClElement.cpp
@@ -327,7 +327,7 @@ const char *ClElement::getComment() const
  */
 void ClElement::arrangeISODirTable()
 {
-   uint tmp, cur=0;
+   uint16 tmp, cur=0;
    /*
     * remove extensions in this order
     */
